requirements.txt
openai~=0.26.5
python-telegram-bot[job-queue, webhooks]==20.0.0
